Updated modelling code to separate file with registered models, updated test and added CoreShellBicelle
models.py now contains the information on registered models including test parameters. The tests have been updated to take advantage of this and will now run through the full set of registered models. Starting parameters have been set for fits so that the test fits converge to the correct values.

The CoreShellBicelle model has been added to the set of registered models.